+++ Changelog.md
@@ -2,10 +2,15 @@
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).   
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+# [1.2.0] - 2022-02-06
+### Added
+- --loc option: Sort AUTHOR names by amount of modified LOC (descendig order, default: alphabetical order)
+### Cahnges
+- Changed AUTHORS not to include bots.
 
 # [1.1.0] - 2022-02-06
 ### Added
-- --comit option: sort AUTHOR names by descending order of the commits 
+- --comit option: Sort AUTHOR names by amount of commits (descendig order, default: alphabetical order)
 
 # [1.0.0] - 2022-01-15
 ### Added